+++ includes/class-max-marine-product-categories-enhancements.php
@@ -159,10 +159,13 @@ class Max_Marine_Product_Categories_Enhancements {
 		$plugin_admin = new Max_Marine_Product_Categories_Enhancements_Admin();
 
 		$this->loader->add_action( 'admin_enqueue_scripts', $plugin_admin, 'enqueue_styles' );
+		$this->loader->add_action( 'admin_enqueue_scripts', $plugin_admin, 'enqueue_scripts' );
 
 		$this->loader->add_filter( 'get_terms', $plugin_admin, 'hide_legacy_categories_from_product_checklist', 10, 3 );
 
 		$this->loader->add_action( 'woocommerce_product_duplicate', $plugin_admin, 'woocommerce_product_duplicate', 999, 2 );
+
+		$this->loader->add_action( 'edit_form_after_editor', $plugin_admin, 'edit_product_page_react_root', 999 );
 	}
 
 	/**