+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * ENH: New function max_marine_product_categories_enhancements_filter_legacy_categories_from_term_ids.
 * ENH: New function max_marine_product_categories_enhancements_filter_legacy_categories_from_terms.
 * ENH: Remove edit term caps from non admin users for the `product_cat` taxonomy.
+* ENH: Remove all current legacy categories when a product is duplicated.
 
 ## [1.0.0] - 2024-12-31
 * Initial release.